python-recorder/requirements.txt
numpy
tk
pyserial